[REFACTOR]: 가격과 관련된 로직들 price.js 파일로 분리
